Conflicts:
	app/controllers/concerns/abstract_mooc_provider_connector.rb
	app/controllers/users_controller.rb
	config/routes.rb

+++ app/workers/abstract_course_worker.rb
@@ -1,3 +1,4 @@
+# -*- encoding : utf-8 -*-
 class AbstractCourseWorker
   include Sidekiq::Worker
   require 'rest_client'
@@ -7,42 +8,37 @@ class AbstractCourseWorker
   end
 
   def load_courses
-    begin
-      response_data = get_course_data
-    rescue SocketError, RestClient::ResourceNotFound, RestClient::SSLCertificateNotVerified => e
-      Rails.logger.error "#{e.class.to_s}: #{e.message}"
-    else
-      handle_response_data response_data
-    end
+    response_data = course_data
+  rescue SocketError, RestClient::ResourceNotFound, RestClient::SSLCertificateNotVerified => e
+    Rails.logger.error "#{e.class}: #{e.message}"
+  else
+    handle_response_data response_data
   end
 
   def mooc_provider
     raise NotImplementedError
   end
 
-  def get_course_data
+  def course_data
     raise NotImplementedError
   end
 
-  def handle_response_data response_data
+  def handle_response_data(_response_data)
     raise NotImplementedError
   end
 
-  def create_update_map mooc_provider
-    update_map = Hash.new
-    Course.where(:mooc_provider_id => mooc_provider.id).each do |course|
+  def create_update_map(mooc_provider)
+    update_map = {}
+    Course.where(mooc_provider_id: mooc_provider.id).each do |course|
       update_map.store(course.id, false)
     end
-    return update_map
+    update_map
   end
 
-  def evaluate_update_map update_map
-    update_map.each do |course_id,updated|
+  def evaluate_update_map(update_map)
+    update_map.each do |course_id, updated|
       course = Course.find(course_id)
-      if !updated and course.present?
-        course.destroy
-      end
+      course.destroy if !updated && course.present?
     end
   end
-
 end
